REFACTOR.패키지 명 topster_album -> topsterALbum
