+++ src/main/java/com/sparta/topster/domain/openApi/service/maniadb/ManiadbService.java
@@ -44,7 +44,7 @@ public class ManiadbService implements OpenApiService {
         URI uri = UriComponentsBuilder.fromUriString("http://www.maniadb.com")
                 .path("api/search/" + query + "/")
                 .queryParam("sr", "album")
-                .queryParam("display", 20)
+                .queryParam("display", 30)
                 .queryParam("v", 0.5)
                 .encode()
                 .build()
